test(pty): give the copy-pasted PTY harness one home
`open_pty`, `drain`, `strip_ansi`, `norm`, `seen`, `recent` and
`wait_for_marker` were pasted into four test files. That duplication is why
#92 could be half-fixed: the cumulative-sink defect was corrected in
tui_key_list_process.rs and stayed open in its three copies, and every future
fix would have had the same shape. 403 duplicated lines are gone.

`open_pty` now takes the size instead of hardcoding one. That was not
cosmetic: the key-list test hid the #71 scroll defect precisely because it
was pinned at 40 rows, where the list fits. A size a test chooses is part of
its contract, so it is an argument. `force_repaint` became `resize_pty`,
which is what it always was.

`recent` settles on one tail length (the three copies used 600, 800 and 900
arbitrarily). Diagnostics only.

Five PTY targets green, fmt and clippy --all-targets --all-features clean,
nothing left in the temp directory.
