+++ tests/tui_planning_reentry_process.rs
@@ -23,14 +23,15 @@
 #![cfg(unix)]
 
 use std::fs::{self, File};
-use std::io::{ErrorKind, Read, Write};
-use std::os::fd::{AsRawFd, FromRawFd};
+use std::io::Write;
 use std::os::unix::fs::PermissionsExt;
 use std::path::{Path, PathBuf};
 use std::process::{Command, Stdio};
 use std::time::{Duration, Instant, SystemTime, UNIX_EPOCH};
 
 mod common;
+
+use common::{drain, open_pty, recent, seen, wait_for_marker};
 
 fn test_root() -> PathBuf {
     let nonce = SystemTime::now()
@@ -90,123 +91,6 @@ fn write_planner(path: &Path) {
     fs::set_permissions(path, permissions).unwrap();
 }
 
-fn open_pty() -> (File, File) {
-    let mut master = -1;
-    let mut slave = -1;
-    let mut size = libc::winsize {
-        ws_row: 40,
-        ws_col: 140,
-        ws_xpixel: 0,
-        ws_ypixel: 0,
-    };
-    let rc = unsafe {
-        libc::openpty(
-            &mut master,
-            &mut slave,
-            std::ptr::null_mut(),
-            std::ptr::null_mut(),
-            &mut size as *mut libc::winsize,
-        )
-    };
-    assert_eq!(rc, 0, "openpty failed: {}", std::io::Error::last_os_error());
-    let master = unsafe { File::from_raw_fd(master) };
-    let slave = unsafe { File::from_raw_fd(slave) };
-    let flags = unsafe { libc::fcntl(master.as_raw_fd(), libc::F_GETFL) };
-    assert!(flags >= 0);
-    let rc = unsafe { libc::fcntl(master.as_raw_fd(), libc::F_SETFL, flags | libc::O_NONBLOCK) };
-    assert_eq!(rc, 0);
-    (master, slave)
-}
-
-/// Drain everything currently readable. Draining continuously is mandatory: a
-/// full PTY buffer blocks the child on its next render.
-fn drain(master: &mut File, sink: &mut Vec<u8>) {
-    let mut buffer = [0_u8; 8192];
-    loop {
-        match master.read(&mut buffer) {
-            Ok(0) => break,
-            Ok(count) => sink.extend_from_slice(&buffer[..count]),
-            Err(error) if error.kind() == ErrorKind::WouldBlock => break,
-            Err(error) => panic!("PTY read failed: {error}"),
-        }
-    }
-}
-
-/// Ratatui positions every wide (CJK) cell with its own cursor-move escape, so
-/// a Korean word is not a contiguous byte run until the escapes are stripped.
-fn strip_ansi(bytes: &[u8]) -> String {
-    let mut out: Vec<u8> = Vec::with_capacity(bytes.len());
-    let mut i = 0;
-    while i < bytes.len() {
-        if bytes[i] == 0x1b {
-            match bytes.get(i + 1) {
-                Some(b'[') => {
-                    i += 2;
-                    while i < bytes.len() && !(0x40..=0x7e).contains(&bytes[i]) {
-                        i += 1;
-                    }
-                    i += usize::from(i < bytes.len());
-                }
-                Some(b']') => {
-                    i += 2;
-                    while i < bytes.len() {
-                        if bytes[i] == 0x07 {
-                            i += 1;
-                            break;
-                        }
-                        if bytes[i] == 0x1b && bytes.get(i + 1) == Some(&b'\\') {
-                            i += 2;
-                            break;
-                        }
-                        i += 1;
-                    }
-                }
-                Some(_) => i += 2,
-                None => break,
-            }
-        } else {
-            out.push(bytes[i]);
-            i += 1;
-        }
-    }
-    String::from_utf8_lossy(&out).into_owned()
-}
-
-/// Whitespace between cells is emitted as cursor motion too, so compare the
-/// visible text with all whitespace removed.
-fn norm(s: &str) -> String {
-    s.chars().filter(|c| !c.is_whitespace()).collect()
-}
-
-fn seen(sink: &[u8], marker: &str) -> bool {
-    norm(&strip_ansi(sink)).contains(&norm(marker))
-}
-
-fn recent(sink: &[u8]) -> String {
-    let text = strip_ansi(sink);
-    let chars: Vec<char> = text.chars().collect();
-    let start = chars.len().saturating_sub(800);
-    chars[start..].iter().collect()
-}
-
-fn wait_for_marker(master: &mut File, sink: &mut Vec<u8>, marker: &str, within: Duration) {
-    let deadline = Instant::now() + within;
-    loop {
-        drain(master, sink);
-        if seen(sink, marker) {
-            return;
-        }
-        assert!(
-            Instant::now() < deadline,
-            "marker {marker:?} not seen within {within:?}; recent output:\n{}",
-            recent(sink)
-        );
-        std::thread::sleep(Duration::from_millis(10));
-    }
-}
-
-/// Poll a workspace predicate while still draining the PTY, so the child never
-/// blocks on render while we wait for its side effect.
 fn wait_until(
     master: &mut File,
     sink: &mut Vec<u8>,
@@ -230,7 +114,7 @@ fn wait_until(
 }
 
 fn spawn_tui(binary: &Path, root: &Path) -> (common::ChildGuard, File, Vec<u8>) {
-    let (master, slave) = open_pty();
+    let (master, slave) = open_pty(40, 140);
     let stdin = Stdio::from(slave.try_clone().unwrap());
     let stdout = Stdio::from(slave.try_clone().unwrap());
     let stderr = Stdio::from(slave);
