+++ tests/tui_startup_process.rs
@@ -1,14 +1,15 @@
 #![cfg(unix)]
 
-use std::fs::{self, File};
+use std::fs;
 use std::io::{ErrorKind, Read, Write};
-use std::os::fd::{AsRawFd, FromRawFd};
 use std::os::unix::fs::PermissionsExt;
 use std::path::{Path, PathBuf};
 use std::process::{Command, Stdio};
 use std::time::{Duration, Instant};
 
 mod common;
+
+use common::open_pty;
 
 fn test_root(name: &str) -> PathBuf {
     std::env::temp_dir().join(format!("yard-tui-startup-{name}-{}", std::process::id()))
@@ -32,34 +33,6 @@ fn write_worker(path: &Path, sentinel: &Path) {
     let mut permissions = fs::metadata(path).unwrap().permissions();
     permissions.set_mode(0o755);
     fs::set_permissions(path, permissions).unwrap();
-}
-
-fn open_pty() -> (File, File) {
-    let mut master = -1;
-    let mut slave = -1;
-    let mut size = libc::winsize {
-        ws_row: 30,
-        ws_col: 120,
-        ws_xpixel: 0,
-        ws_ypixel: 0,
-    };
-    let rc = unsafe {
-        libc::openpty(
-            &mut master,
-            &mut slave,
-            std::ptr::null_mut(),
-            std::ptr::null_mut(),
-            &mut size as *mut libc::winsize,
-        )
-    };
-    assert_eq!(rc, 0, "openpty failed: {}", std::io::Error::last_os_error());
-    let master = unsafe { File::from_raw_fd(master) };
-    let slave = unsafe { File::from_raw_fd(slave) };
-    let flags = unsafe { libc::fcntl(master.as_raw_fd(), libc::F_GETFL) };
-    assert!(flags >= 0);
-    let rc = unsafe { libc::fcntl(master.as_raw_fd(), libc::F_SETFL, flags | libc::O_NONBLOCK) };
-    assert_eq!(rc, 0);
-    (master, slave)
 }
 
 #[test]
@@ -92,7 +65,7 @@ fn slow_probe_and_recovery_do_not_block_first_safe_tui_frame() {
     write_worker(&bin_dir.join("codex"), &sentinel);
     write_worker(&bin_dir.join("claude"), &sentinel);
 
-    let (mut master, slave) = open_pty();
+    let (mut master, slave) = open_pty(30, 120);
     let stdin = Stdio::from(slave.try_clone().unwrap());
     let stdout = Stdio::from(slave.try_clone().unwrap());
     let stderr = Stdio::from(slave);
